Extract ServerClaim scheduling into a dedicated scheduler controller
