+++ config/rbac/role.yaml
@@ -43,7 +43,6 @@ rules:
   - bmcusers
   - endpoints
   - serverbootconfigurations
-  - serverclaims
   - serverconfigurations
   - serverreadinessrules
   - servers
@@ -83,3 +82,13 @@ rules:
   - get
   - patch
   - update
+- apiGroups:
+  - metal.ironcore.dev
+  resources:
+  - serverclaims
+  verbs:
+  - get
+  - list
+  - patch
+  - update
+  - watch
