+++ internal/controller/serverclaim_scheduler_controller_test.go
@@ -1,0 +1,684 @@
+// SPDX-FileCopyrightText: 2026 SAP SE or an SAP affiliate company and IronCore contributors
+// SPDX-License-Identifier: Apache-2.0
+
+package controller
+
+import (
+	"context"
+
+	metalv1alpha1 "github.com/ironcore-dev/metal-operator/api/v1alpha1"
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+	v1 "k8s.io/api/core/v1"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/labels"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	. "sigs.k8s.io/controller-runtime/pkg/envtest/komega"
+)
+
+var _ = Describe("ServerClaim Scheduler Controller", func() {
+	ns := SetupTest(nil)
+
+	var (
+		server    *metalv1alpha1.Server
+		bmcSecret *metalv1alpha1.BMCSecret
+	)
+
+	BeforeEach(func(ctx SpecContext) {
+		By("Creating a BMCSecret")
+		bmcSecret = &metalv1alpha1.BMCSecret{
+			ObjectMeta: metav1.ObjectMeta{
+				GenerateName: "test-",
+			},
+			Data: map[string][]byte{
+				metalv1alpha1.BMCSecretUsernameKeyName: []byte("foo"),
+				metalv1alpha1.BMCSecretPasswordKeyName: []byte("bar"),
+			},
+		}
+		Expect(k8sClient.Create(ctx, bmcSecret)).To(Succeed())
+
+		By("Creating a Server")
+		server = &metalv1alpha1.Server{
+			ObjectMeta: metav1.ObjectMeta{
+				GenerateName: "test-claim-",
+			},
+			Spec: metalv1alpha1.ServerSpec{
+				SystemUUID: "38947555-7742-3448-3784-823347823834",
+				BMC: &metalv1alpha1.BMCAccess{
+					Protocol: metalv1alpha1.Protocol{
+						Name: metalv1alpha1.ProtocolRedfishLocal,
+						Port: MockServerPort,
+					},
+					Address: MockServerIP,
+					BMCSecretRef: v1.LocalObjectReference{
+						Name: bmcSecret.Name,
+					},
+				},
+			},
+		}
+		Expect(k8sClient.Create(ctx, server)).To(Succeed())
+
+		Eventually(UpdateStatus(server, func() {
+			server.Status.State = metalv1alpha1.ServerStateAvailable
+		})).Should(Succeed())
+	})
+
+	AfterEach(func(ctx SpecContext) {
+		Expect(k8sClient.Delete(ctx, server)).To(Succeed())
+		Expect(k8sClient.Delete(ctx, bmcSecret)).To(Succeed())
+		EnsureCleanState(ctx)
+	})
+
+	It("should successfully claim a server by reference and label selector", func(ctx SpecContext) {
+		By("Patching Server labels")
+		Eventually(Update(server, func() {
+			server.Labels = map[string]string{
+				"type": "storage",
+				"env":  "staging",
+			}
+		})).Should(Succeed())
+
+		By("Creating a ServerClaim")
+		claim := &metalv1alpha1.ServerClaim{
+			ObjectMeta: metav1.ObjectMeta{
+				Namespace:    ns.Name,
+				GenerateName: "test-",
+			},
+			Spec: metalv1alpha1.ServerClaimSpec{
+				Power:     metalv1alpha1.PowerOff,
+				ServerRef: &v1.LocalObjectReference{Name: server.Name},
+				ServerSelector: &metav1.LabelSelector{
+					MatchLabels: map[string]string{"type": "storage"},
+					MatchExpressions: []metav1.LabelSelectorRequirement{{
+						Key:      "env",
+						Operator: metav1.LabelSelectorOpIn,
+						Values:   []string{"test", "staging"},
+					}},
+				},
+				Image: "foo:bar",
+			},
+		}
+		Expect(k8sClient.Create(ctx, claim)).To(Succeed())
+
+		By("Ensuring that the Server has the correct claim ref")
+		Eventually(Object(server)).Should(SatisfyAll(
+			HaveField("Spec.ServerClaimRef.Name", claim.Name),
+			HaveField("Status.State", metalv1alpha1.ServerStateReserved),
+		))
+
+		By("Ensuring that the ServerClaim is bound")
+		Eventually(Object(claim)).Should(SatisfyAll(
+			HaveField("Finalizers", ContainElement(serverClaimFinalizer)),
+			HaveField("Status.Phase", metalv1alpha1.PhaseBound),
+			HaveField("Spec.ServerRef", Not(BeNil())),
+			HaveField("Spec.ServerRef.Name", server.Name),
+		))
+
+		By("Deleting the ServerClaim")
+		Expect(k8sClient.Delete(ctx, claim)).To(Succeed())
+
+		By("Ensuring that the ServerClaim is deleted")
+		Eventually(Get(claim)).Should(Satisfy(apierrors.IsNotFound))
+
+		By("Ensuring that the Server is available")
+		Eventually(Object(server)).Should(SatisfyAll(
+			HaveField("Spec.ServerClaimRef", BeNil()),
+			HaveField("Spec.BootConfigurationRef", BeNil()),
+			HaveField("Status.State", metalv1alpha1.ServerStateAvailable),
+		))
+	})
+
+	It("should successfully claim a server by label selector", func(ctx SpecContext) {
+		By("Patching Server labels")
+		Eventually(Update(server, func() {
+			server.Labels = map[string]string{
+				"type": "storage",
+				"env":  "prod",
+			}
+		})).Should(Succeed())
+
+		By("Creating a ServerClaim")
+		claim := &metalv1alpha1.ServerClaim{
+			ObjectMeta: metav1.ObjectMeta{
+				Namespace:    ns.Name,
+				GenerateName: "test-",
+			},
+			Spec: metalv1alpha1.ServerClaimSpec{
+				Power: metalv1alpha1.PowerOff,
+				ServerSelector: &metav1.LabelSelector{
+					MatchLabels: map[string]string{"type": "storage"},
+					MatchExpressions: []metav1.LabelSelectorRequirement{{
+						Key:      "env",
+						Operator: metav1.LabelSelectorOpIn,
+						Values:   []string{"prod"},
+					}},
+				},
+				Image: "foo:bar",
+			},
+		}
+		Expect(k8sClient.Create(ctx, claim)).To(Succeed())
+
+		By("Ensuring that the ServerClaim is bound")
+		Eventually(Object(claim)).Should(SatisfyAll(
+			HaveField("Finalizers", ContainElement(serverClaimFinalizer)),
+			HaveField("Spec.ServerRef", Equal(&v1.LocalObjectReference{Name: server.Name})),
+			HaveField("Status.Phase", metalv1alpha1.PhaseBound),
+		))
+
+		By("Ensuring that the Server has the correct claim ref")
+		Eventually(Object(server)).Should(SatisfyAll(
+			HaveField("Spec.ServerClaimRef", &metalv1alpha1.ImmutableObjectReference{
+				Name:      claim.Name,
+				Namespace: claim.Namespace,
+			}),
+			HaveField("Status.State", metalv1alpha1.ServerStateReserved),
+			HaveField("Status.PowerState", metalv1alpha1.ServerOffPowerState),
+		))
+
+		By("Deleting the ServerClaim")
+		Expect(k8sClient.Delete(ctx, claim)).To(Succeed())
+
+		By("Ensuring that the ServerClaim is deleted")
+		Eventually(Get(claim)).Should(Satisfy(apierrors.IsNotFound))
+
+		By("Ensuring that the Server is available")
+		Eventually(Object(server)).Should(SatisfyAll(
+			HaveField("Spec.ServerClaimRef", BeNil()),
+			HaveField("Spec.BootConfigurationRef", BeNil()),
+			HaveField("Status.State", metalv1alpha1.ServerStateAvailable),
+			HaveField("Status.PowerState", metalv1alpha1.ServerOffPowerState),
+		))
+	})
+
+	It("should not claim a server in a non-available state", func(ctx SpecContext) {
+		By("Patching the Server to Initial state")
+		Eventually(UpdateStatus(server, func() {
+			server.Status.State = metalv1alpha1.ServerStateInitial
+		})).Should(Succeed())
+
+		By("Creating a ServerClaim")
+		claim := &metalv1alpha1.ServerClaim{
+			ObjectMeta: metav1.ObjectMeta{
+				Namespace:    ns.Name,
+				GenerateName: "test-",
+			},
+			Spec: metalv1alpha1.ServerClaimSpec{
+				Power:     metalv1alpha1.PowerOn,
+				ServerRef: &v1.LocalObjectReference{Name: server.Name},
+				Image:     "foo:bar",
+			},
+		}
+		Expect(k8sClient.Create(ctx, claim)).To(Succeed())
+
+		By("Ensuring that the Server has no claim ref")
+		Eventually(Object(server)).Should(SatisfyAll(
+			HaveField("Spec.ServerClaimRef", BeNil()),
+			HaveField("Status.State", metalv1alpha1.ServerStateInitial),
+		))
+
+		By("Ensuring that the ServerClaim stays unbound")
+		Eventually(Object(claim)).Should(SatisfyAll(
+			HaveField("Finalizers", ContainElement(serverClaimFinalizer)),
+			HaveField("Status.Phase", metalv1alpha1.PhaseUnbound),
+		))
+
+		By("Ensuring that the ServerBootConfiguration has not been created")
+		config := &metalv1alpha1.ServerBootConfiguration{
+			ObjectMeta: metav1.ObjectMeta{
+				Namespace: ns.Name,
+				Name:      claim.Name,
+			},
+		}
+		Eventually(Get(config)).Should(Satisfy(apierrors.IsNotFound))
+
+		By("Removing the ServerClaim")
+		Expect(k8sClient.Delete(ctx, claim)).To(Succeed())
+
+		By("Ensuring that the ServerClaim is deleted")
+		Eventually(Get(claim)).Should(Satisfy(apierrors.IsNotFound))
+	})
+
+	It("should not claim a server with set claim ref", func(ctx SpecContext) {
+		By("Updating the Server to available state")
+		Eventually(UpdateStatus(server, func() {
+			server.Status.State = metalv1alpha1.ServerStateAvailable
+		})).Should(Succeed())
+
+		By("Creating a ServerClaim")
+		claim := &metalv1alpha1.ServerClaim{
+			ObjectMeta: metav1.ObjectMeta{
+				Namespace:    ns.Name,
+				GenerateName: "test-",
+			},
+			Spec: metalv1alpha1.ServerClaimSpec{
+				Power:     metalv1alpha1.PowerOn,
+				ServerRef: &v1.LocalObjectReference{Name: server.Name},
+				Image:     "foo:bar",
+			},
+		}
+		Expect(k8sClient.Create(ctx, claim)).To(Succeed())
+
+		By("Ensuring that the Server has claim ref")
+		Eventually(Object(server)).Should(SatisfyAll(
+			HaveField("Spec.ServerClaimRef", &metalv1alpha1.ImmutableObjectReference{
+				Namespace: ns.Name,
+				Name:      claim.Name,
+			}),
+			HaveField("Status.State", metalv1alpha1.ServerStateReserved),
+		))
+
+		By("Ensuring that the ServerClaim is bound")
+		Eventually(Object(claim)).Should(SatisfyAll(
+			HaveField("Finalizers", ContainElement(serverClaimFinalizer)),
+			HaveField("Status.Phase", metalv1alpha1.PhaseBound),
+		))
+
+		By("Creating another ServerClaim")
+		claim2 := &metalv1alpha1.ServerClaim{
+			ObjectMeta: metav1.ObjectMeta{
+				Namespace:    ns.Name,
+				GenerateName: "test-duplicate-",
+			},
+			Spec: metalv1alpha1.ServerClaimSpec{
+				Power:     metalv1alpha1.PowerOn,
+				ServerRef: &v1.LocalObjectReference{Name: server.Name},
+				Image:     "foo:bar2",
+			},
+		}
+		Expect(k8sClient.Create(ctx, claim2)).To(Succeed())
+
+		By("Ensuring that the ServerClaim is Unbound")
+		Eventually(Object(claim2)).Should(SatisfyAll(
+			HaveField("Finalizers", ContainElement(serverClaimFinalizer)),
+			HaveField("Status.Phase", metalv1alpha1.PhaseUnbound),
+		))
+
+		By("Ensuring that the Server has claim ref")
+		Eventually(Object(server)).Should(SatisfyAll(
+			HaveField("Spec.ServerClaimRef", &metalv1alpha1.ImmutableObjectReference{
+				Namespace: ns.Name,
+				Name:      claim.Name,
+			}),
+			HaveField("Status.State", metalv1alpha1.ServerStateReserved),
+		))
+
+		By("Ensuring that the ServerBootConfiguration has not been created")
+		config := &metalv1alpha1.ServerBootConfiguration{
+			ObjectMeta: metav1.ObjectMeta{
+				Namespace: ns.Name,
+				Name:      claim2.Name,
+			},
+		}
+		Eventually(Get(config)).Should(Satisfy(apierrors.IsNotFound))
+
+		By("Removing the ServerClaim")
+		Expect(k8sClient.Delete(ctx, claim)).To(Succeed())
+		By("Removing the ServerClaim2")
+		Expect(k8sClient.Delete(ctx, claim2)).To(Succeed())
+
+		By("Ensuring that the ServerClaim is deleted")
+		Eventually(Get(claim)).Should(Satisfy(apierrors.IsNotFound))
+		Eventually(Get(claim2)).Should(Satisfy(apierrors.IsNotFound))
+
+	})
+
+	It("should not claim a server when labels do not match selector", func(ctx SpecContext) {
+		By("Creating a ServerClaim")
+		claim := &metalv1alpha1.ServerClaim{
+			ObjectMeta: metav1.ObjectMeta{
+				Namespace:    ns.Name,
+				GenerateName: "test-",
+			},
+			Spec: metalv1alpha1.ServerClaimSpec{
+				Power:     metalv1alpha1.PowerOn,
+				ServerRef: &v1.LocalObjectReference{Name: server.Name},
+				ServerSelector: &metav1.LabelSelector{
+					MatchLabels: map[string]string{"type": "storage"},
+					MatchExpressions: []metav1.LabelSelectorRequirement{{
+						Key:      "env",
+						Operator: metav1.LabelSelectorOpIn,
+						Values:   []string{"test", "staging"},
+					}},
+				},
+				Image: "foo:bar",
+			},
+		}
+		Expect(k8sClient.Create(ctx, claim)).To(Succeed())
+
+		By("Patching the Server to available state")
+		Eventually(UpdateStatus(server, func() {
+			server.Status.State = metalv1alpha1.ServerStateAvailable
+		})).Should(Succeed())
+
+		By("Ensuring that the Server has no claim ref")
+		Eventually(Object(server)).Should(SatisfyAll(
+			HaveField("Spec.ServerClaimRef", BeNil()),
+		))
+
+		By("Ensuring that the ServerClaim is unbound")
+		Eventually(Object(claim)).Should(SatisfyAll(
+			HaveField("Finalizers", ContainElement(serverClaimFinalizer)),
+			HaveField("Status.Phase", metalv1alpha1.PhaseUnbound),
+		))
+
+		By("Removing the ServerClaim")
+		Expect(k8sClient.Delete(ctx, claim)).To(Succeed())
+
+		By("Ensuring that the ServerClaim is deleted")
+		Eventually(Get(claim)).Should(Satisfy(apierrors.IsNotFound))
+	})
+
+	It("should not claim a server whose taints are not tolerated by the claim", func(ctx SpecContext) {
+		By("Adding a NoBind taint to the server")
+		Eventually(Update(server, func() {
+			server.Spec.Taints = []metalv1alpha1.Taint{
+				{Key: "dedicated", Value: "gpu", Effect: metalv1alpha1.TaintEffectNoBind},
+			}
+		})).Should(Succeed())
+
+		By("Creating a ServerClaim without tolerations")
+		claim := &metalv1alpha1.ServerClaim{
+			ObjectMeta: metav1.ObjectMeta{
+				Namespace:    ns.Name,
+				GenerateName: "test-taint-",
+			},
+			Spec: metalv1alpha1.ServerClaimSpec{
+				Power:     metalv1alpha1.PowerOn,
+				ServerRef: &v1.LocalObjectReference{Name: server.Name},
+				Image:     "foo:bar",
+			},
+		}
+		Expect(k8sClient.Create(ctx, claim)).To(Succeed())
+
+		By("Ensuring the server is not claimed")
+		Consistently(Object(server)).Should(
+			HaveField("Spec.ServerClaimRef", BeNil()),
+		)
+		Consistently(Object(claim)).Should(
+			HaveField("Status.Phase", Not(Equal(metalv1alpha1.PhaseBound))),
+		)
+
+		By("Removing the ServerClaim")
+		Expect(k8sClient.Delete(ctx, claim)).To(Succeed())
+		Eventually(Get(claim)).Should(Satisfy(apierrors.IsNotFound))
+	})
+
+	It("should claim a server with a NoBind taint when the claim has a matching toleration", func(ctx SpecContext) {
+		By("Adding a NoBind taint to the server")
+		Eventually(Update(server, func() {
+			server.Spec.Taints = []metalv1alpha1.Taint{
+				{Key: "dedicated", Value: "gpu", Effect: metalv1alpha1.TaintEffectNoBind},
+			}
+		})).Should(Succeed())
+
+		By("Creating a ServerClaim with a matching Exists toleration")
+		claim := &metalv1alpha1.ServerClaim{
+			ObjectMeta: metav1.ObjectMeta{
+				Namespace:    ns.Name,
+				GenerateName: "test-toleration-",
+			},
+			Spec: metalv1alpha1.ServerClaimSpec{
+				Power:     metalv1alpha1.PowerOn,
+				ServerRef: &v1.LocalObjectReference{Name: server.Name},
+				Image:     "foo:bar",
+				Tolerations: []metalv1alpha1.Toleration{
+					{Key: "dedicated", Operator: metalv1alpha1.TolerationOperatorExists},
+				},
+			},
+		}
+		Expect(k8sClient.Create(ctx, claim)).To(Succeed())
+
+		By("Ensuring the server is claimed")
+		Eventually(Object(claim)).Should(
+			HaveField("Status.Phase", Equal(metalv1alpha1.PhaseBound)),
+		)
+
+		By("Removing the ServerClaim")
+		Expect(k8sClient.Delete(ctx, claim)).To(Succeed())
+		Eventually(Get(claim)).Should(Satisfy(apierrors.IsNotFound))
+	})
+
+	It("should default an empty taint effect to NoBind", func(ctx SpecContext) {
+		By("Updating the server with a taint that has no effect set")
+		Eventually(Update(server, func() {
+			server.Spec.Taints = []metalv1alpha1.Taint{
+				{Key: "dedicated"},
+			}
+		})).Should(Succeed())
+
+		By("Verifying the taint effect is defaulted to NoBind")
+		Eventually(Object(server)).Should(
+			HaveField("Spec.Taints", ConsistOf(
+				HaveField("Effect", metalv1alpha1.TaintEffectNoBind),
+			)),
+		)
+	})
+
+	It("should not claim a cordoned server by explicit ServerRef", func(ctx SpecContext) {
+		By("Cordoning the server")
+		Eventually(Update(server, func() {
+			server.Spec.Unclaimable = true
+		})).Should(Succeed())
+
+		By("Creating a ServerClaim referencing the cordoned server")
+		claim := &metalv1alpha1.ServerClaim{
+			ObjectMeta: metav1.ObjectMeta{
+				Namespace:    ns.Name,
+				GenerateName: "test-cordon-",
+			},
+			Spec: metalv1alpha1.ServerClaimSpec{
+				Power:     metalv1alpha1.PowerOn,
+				ServerRef: &v1.LocalObjectReference{Name: server.Name},
+				Image:     "foo:bar",
+			},
+		}
+		Expect(k8sClient.Create(ctx, claim)).To(Succeed())
+
+		By("Ensuring the cordoned server is not claimed")
+		Consistently(Object(server)).Should(
+			HaveField("Spec.ServerClaimRef", BeNil()),
+		)
+
+		By("Ensuring the claim is registered and stays unbound")
+		Eventually(Object(claim)).Should(SatisfyAll(
+			HaveField("Finalizers", ContainElement(serverClaimFinalizer)),
+			HaveField("Status.Phase", Equal(metalv1alpha1.PhaseUnbound)),
+		))
+		Consistently(Object(claim)).Should(
+			HaveField("Status.Phase", Equal(metalv1alpha1.PhaseUnbound)),
+		)
+
+		By("Uncordoning the server")
+		Eventually(Update(server, func() {
+			server.Spec.Unclaimable = false
+		})).Should(Succeed())
+
+		By("Ensuring the server is now claimed")
+		Eventually(Object(server)).Should(SatisfyAll(
+			HaveField("Spec.ServerClaimRef.Name", claim.Name),
+			HaveField("Status.State", metalv1alpha1.ServerStateReserved),
+		))
+		Eventually(Object(claim)).Should(
+			HaveField("Status.Phase", Equal(metalv1alpha1.PhaseBound)),
+		)
+
+		By("Removing the ServerClaim")
+		Expect(k8sClient.Delete(ctx, claim)).To(Succeed())
+		Eventually(Get(claim)).Should(Satisfy(apierrors.IsNotFound))
+	})
+
+	It("should skip a cordoned server when claiming by selector", func(ctx SpecContext) {
+		By("Labeling and cordoning the server")
+		Eventually(Update(server, func() {
+			server.Labels = map[string]string{"type": "storage"}
+			server.Spec.Unclaimable = true
+		})).Should(Succeed())
+
+		By("Creating a ServerClaim with a matching selector")
+		claim := &metalv1alpha1.ServerClaim{
+			ObjectMeta: metav1.ObjectMeta{
+				Namespace:    ns.Name,
+				GenerateName: "test-cordon-selector-",
+			},
+			Spec: metalv1alpha1.ServerClaimSpec{
+				Power: metalv1alpha1.PowerOff,
+				ServerSelector: &metav1.LabelSelector{
+					MatchLabels: map[string]string{"type": "storage"},
+				},
+				Image: "foo:bar",
+			},
+		}
+		Expect(k8sClient.Create(ctx, claim)).To(Succeed())
+
+		By("Ensuring the cordoned server is not claimed")
+		Consistently(Object(server)).Should(
+			HaveField("Spec.ServerClaimRef", BeNil()),
+		)
+
+		By("Ensuring the claim is registered and stays unbound")
+		Eventually(Object(claim)).Should(SatisfyAll(
+			HaveField("Finalizers", BeEmpty()),
+			HaveField("Status.Phase", Equal(metalv1alpha1.PhaseUnbound)),
+		))
+		Consistently(Object(claim)).Should(
+			HaveField("Status.Phase", Equal(metalv1alpha1.PhaseUnbound)),
+		)
+
+		By("Removing the ServerClaim")
+		Expect(k8sClient.Delete(ctx, claim)).To(Succeed())
+		Eventually(Get(claim)).Should(Satisfy(apierrors.IsNotFound))
+	})
+})
+
+var _ = Describe("Server Claiming", MustPassRepeatedly(5), func() {
+	ns := SetupTest(nil)
+
+	makeServer := func(ctx context.Context) {
+		server := metalv1alpha1.Server{
+			ObjectMeta: metav1.ObjectMeta{
+				GenerateName: "server-",
+				Annotations: map[string]string{
+					metalv1alpha1.OperationAnnotation: metalv1alpha1.OperationAnnotationIgnore,
+				},
+				Labels: map[string]string{"foo": "bar"},
+			},
+			Spec: metalv1alpha1.ServerSpec{
+				SystemUUID: "38947555-7742-3448-3784-823347823834",
+			},
+		}
+		ExpectWithOffset(1, k8sClient.Create(ctx, &server)).To(Succeed())
+		server.Status.State = metalv1alpha1.ServerStateAvailable
+		server.Status.PowerState = metalv1alpha1.ServerOffPowerState
+		ExpectWithOffset(1, k8sClient.Status().Update(ctx, &server)).To(Succeed())
+	}
+
+	makeClaim := func(ctx context.Context, labelSelector *metav1.LabelSelector) {
+		claim := metalv1alpha1.ServerClaim{
+			ObjectMeta: metav1.ObjectMeta{
+				GenerateName: "claim-",
+				Namespace:    ns.Name,
+			},
+			Spec: metalv1alpha1.ServerClaimSpec{
+				ServerSelector: labelSelector,
+			},
+		}
+		ExpectWithOffset(1, k8sClient.Create(ctx, &claim)).To(Succeed())
+	}
+
+	countUniqueBoundServers := func(ctx context.Context, serverCount int) func(Gomega) int {
+		return func(g Gomega) int {
+			var serverList metalv1alpha1.ServerList
+			g.Expect(k8sClient.List(ctx, &serverList, client.MatchingLabels{"foo": "bar"})).To(Succeed())
+			g.Expect(serverList.Items).To(HaveLen(serverCount))
+			claimNames := make(map[string]struct{})
+			for _, server := range serverList.Items {
+				if server.Spec.ServerClaimRef != nil {
+					claimNames[server.Spec.ServerClaimRef.Name] = struct{}{}
+				}
+			}
+			return len(claimNames)
+		}
+	}
+
+	countUniqueBoundClaims := func(ctx context.Context) func(Gomega) int {
+		return func(g Gomega) int {
+			var claimList metalv1alpha1.ServerClaimList
+			g.Expect(k8sClient.List(ctx, &claimList, client.InNamespace(ns.Name))).To(Succeed())
+			serverNames := make(map[string]struct{})
+			for _, claim := range claimList.Items {
+				if claim.Spec.ServerRef != nil {
+					serverNames[claim.Spec.ServerRef.Name] = struct{}{}
+				}
+			}
+			return len(serverNames)
+		}
+	}
+
+	AfterEach(func(ctx SpecContext) {
+		claimList := &metalv1alpha1.ServerClaimList{}
+		Eventually(List(claimList, client.InNamespace(ns.Name))).Should(Succeed())
+		for _, claim := range claimList.Items {
+			Expect(k8sClient.Delete(ctx, &claim)).To(Succeed())
+		}
+
+		serverList := &metalv1alpha1.ServerList{}
+		Eventually(List(serverList, client.MatchingLabels{
+			"foo": "bar",
+		})).Should(Succeed())
+		for _, server := range serverList.Items {
+			Expect(k8sClient.Delete(ctx, &server)).To(Succeed())
+		}
+		EnsureCleanState(ctx)
+	})
+
+	It("should bind four out of ten servers for four best effort claims", func(ctx SpecContext) {
+		for range 10 {
+			makeServer(ctx)
+		}
+		for range 4 {
+			makeClaim(ctx, nil)
+		}
+		Eventually(countUniqueBoundServers(ctx, 10)).Should(Equal(4))
+		Consistently(countUniqueBoundServers(ctx, 10)).Should(Equal(4))
+		Eventually(countUniqueBoundClaims(ctx)).Should(Equal(4))
+		Consistently(countUniqueBoundClaims(ctx)).Should(Equal(4))
+	})
+
+	It("should bind four out of ten servers for four label selector claims", func(ctx SpecContext) {
+		for range 10 {
+			makeServer(ctx)
+		}
+		for range 4 {
+			makeClaim(ctx, metav1.SetAsLabelSelector(labels.Set{"foo": "bar"}))
+		}
+		Eventually(countUniqueBoundServers(ctx, 10)).Should(Equal(4))
+		Consistently(countUniqueBoundServers(ctx, 10)).Should(Equal(4))
+		Eventually(countUniqueBoundClaims(ctx)).Should(Equal(4))
+		Consistently(countUniqueBoundClaims(ctx)).Should(Equal(4))
+	})
+
+	It("should not bind the same server to multiple best effort claims", func(ctx SpecContext) {
+		By("Creating eight ServerClaims")
+		for range 8 {
+			makeClaim(ctx, nil)
+		}
+		makeServer(ctx)
+		Eventually(countUniqueBoundServers(ctx, 1)).Should(Equal(1))
+		Consistently(countUniqueBoundServers(ctx, 1)).Should(Equal(1))
+		Eventually(countUniqueBoundClaims(ctx)).Should(Equal(1))
+		Consistently(countUniqueBoundClaims(ctx)).Should(Equal(1))
+	})
+
+	It("should not bind the same server to multiple label selector claims", func(ctx SpecContext) {
+		By("Creating eight ServerClaims")
+		for range 8 {
+			makeClaim(ctx, metav1.SetAsLabelSelector(labels.Set{"foo": "bar"}))
+		}
+		makeServer(ctx)
+		Eventually(countUniqueBoundServers(ctx, 1)).Should(Equal(1))
+		Consistently(countUniqueBoundServers(ctx, 1)).Should(Equal(1))
+		Eventually(countUniqueBoundClaims(ctx)).Should(Equal(1))
+		Consistently(countUniqueBoundClaims(ctx)).Should(Equal(1))
+	})
+})
